requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Database
pymongo==4.6.0
motor==3.3.2

# Vector Database
pinecone-client==2.2.4
faiss-cpu==1.7.4

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
aiofiles==23.2.0

# LLM and Embeddings
google-generativeai==0.3.2
sentence-transformers==2.2.2
openai==1.3.7

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
httpx==0.25.2
tenacity==8.2.3
